refactor: delegate safety analysis logic to utils
